% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{bed2GR}
\alias{bed2GR}
\title{Read genomic ranges from bed files (zero based coordinates)}
\usage{
bed2GR(filename, nfields = NULL)
}
\arguments{
\item{filename}{name of the bed file}

\item{nfields}{number of fields in the bed file to be parsed. Default is to
automatically detect the number of fields.}
}
\value{
GRanges object with the coordinates from the bed file
}
\description{
Read genomic ranges from bed files (zero based coordinates)
}

