% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{countBamInGRanges}
\alias{countBamInGRanges}
\title{Count reads from a bam file in genomic ranges}
\usage{
countBamInGRanges(bam.file, granges, min.mapq = NULL, read.width = 1)
}
\arguments{
\item{bam.file}{filename of the bam file. The bam file must be sorted and
indexed}

\item{granges}{GRanges object in which to count the reads}

\item{min.mapq}{minimal mapping quality. Default is to ignore mapping quality}

\item{read.width}{the length of a read to be used for overlaps. Default is to #'        use just the start position of the read.}
}
\value{
count vector of the same length as the GRanges object
}
\description{
Count reads from a bam file in genomic ranges
}

