% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{coverageBamInGRanges}
\alias{coverageBamInGRanges}
\title{Coverage profile from a bam file across genomic ranges}
\usage{
coverageBamInGRanges(bam.file, granges, min.mapq = NULL,
  reads.collapsed = FALSE, width = NULL)
}
\arguments{
\item{bam.file}{filename of the bam file. The bam file must be sorted and
indexed}

\item{granges}{GRanges object in which to count the reads. Note: all ranges
must be of the same width}

\item{min.mapq}{minimal mapping quality. Default is to ignore mapping quality}

\item{reads.collapsed}{logical indicating whether duplicated reads were
removed and counts were added as _x to the readnames (this is the case
for some of the tools from the Rajewski lab)}

\item{width}{length of the sequencing reads, by default this will be extracted
from the bam file}
}
\value{
count matrix with rows for each element of the GRanges object
}
\description{
Coverage profile from a bam file across genomic ranges
}

