% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{extract}
\alias{extract}
\title{Helper function to extract text using regular expressions}
\usage{
extract(pattern, string, perl = TRUE)
}
\arguments{
\item{pattern}{regular expression with grouping}

\item{string}{character vector from which to extract the text}

\item{perl}{logical indicating whether regular expression is perl style}
}
\value{
character vector with the extracted text
}
\description{
Helper function to extract text using regular expressions
}

