% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{get.count.matrix}
\alias{get.count.matrix}
\title{Get a count matrix from bam files that match a certain filename pattern}
\usage{
get.count.matrix(granges, dir, pattern)
}
\arguments{
\item{granges}{GRanges object in which to count reads}

\item{dir}{name of the directory that contains the bam files}

\item{pattern}{regulator expression to identify bam files in the directory}
}
\value{
a count matrix with rows for each element of the GRanges object and
        columns for each bam file
}
\description{
Get a count matrix from bam files that match a certain filename pattern
}

