% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{get.htseq.count.matrix}
\alias{get.htseq.count.matrix}
\title{Get a count matrix from htseq-count files that match a filename pattern}
\usage{
get.htseq.count.matrix(dir, pattern)
}
\arguments{
\item{dir}{name of the directory that contains the bam files}

\item{pattern}{regulator expression to identify bam files in the directory}
}
\value{
a count matrix with rows for each element of the GRanges object and
        columns for each bam file
}
\description{
Get a count matrix from htseq-count files that match a filename pattern
}

