% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{get.rpkm}
\alias{get.rpkm}
\title{Normalize to reads per kilobase per million sequences}
\usage{
get.rpkm(annotation, counts)
}
\arguments{
\item{annotation}{GRanges object with exons that have a "gene_id" field in
the meta data}

\item{counts}{read count matrix per gene. The rownames have to be the same as
those used in the gene_id field.}
}
\value{
matrix of RPKM values
}
\description{
Normalize to reads per kilobase per million sequences
}

