% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{gff2GR}
\alias{gff2GR}
\title{Read gene annotations in gff format}
\usage{
gff2GR(filename, gffAttrNames = NULL)
}
\arguments{
\item{filename}{gff filename}

\item{gffAttrNames}{character vector with attribute names that should be
parsed from the gtf file. Usually these are "ID", "Parent"
and the like. By default no attributes are parsed.}
}
\value{
a GRanges object with all entries from the gff file with metadata
        columns for source, type and the attributes parsed
}
\description{
Read gene annotations in gff format
}

