% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{GR2gff}
\alias{GR2gff}
\title{Export GRanges to gff format}
\usage{
GR2gff(regions, filename, feature.type = "experimental_feature",
  src = "GenomicRanges", score = ".", phase = ".")
}
\arguments{
\item{regions}{GRanges object to be saved}

\item{filename}{name of the gff file}

\item{feature.type}{character for the feature type field}

\item{src}{character for the source type field}

\item{score}{numeric or character for the score field}

\item{phase}{numeric or character for the phase field}
}
\description{
Export GRanges to gff format
}

