% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{GR2gtf}
\alias{GR2gtf}
\title{Export GRanges to gtf format}
\usage{
GR2gtf(regions, filename, feature.type = "experimental_feature",
  src = "GenomicRanges", score = ".", phase = ".", attributes = NULL,
  ...)
}
\arguments{
\item{regions}{GRanges object to be saved}

\item{filename}{name of the gtf file}

\item{feature.type}{character for the feature type field}

\item{src}{character for the source type field}

\item{score}{numeric or character for the score field}

\item{phase}{numeric or character for the phase field}

\item{attributes}{character vector of columns to be saved}

\item{...}{additional arguments passed to \code{\link{write.table}}}
}
\description{
Export GRanges to gtf format
}

