% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{gtf2GR}
\alias{gtf2GR}
\title{Read gene annotations in gtf format}
\usage{
gtf2GR(filename, gtfAttrNames = NULL)
}
\arguments{
\item{filename}{gtf filename}

\item{gtfAttrNames}{character vector with attribute names that should be
parsed from the gtf file. Usually these are "gene_id", "gene_name",
"transcript_id" and the like. By default no attributes are parsed.}
}
\value{
a GRanges object with all entries from the gtf file with metadata
        columns for source, type and the attributes parsed
}
\description{
Read gene annotations in gtf format
}

