% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_peaks.R
\name{macs2GRange}
\alias{macs2GRange}
\title{Transform a MACS output peakfile into a GRanges object}
\usage{
macs2GRange(peaks, chroms = NULL)
}
\arguments{
\item{peaks}{a data.frame with output from MACS}

\item{chroms}{character vector with chromosome names to be used. Default is
to use all chromosome names from the MACS file}
}
\value{
a GRanges object with the MACS peak calls
}
\description{
Transform a MACS output peakfile into a GRanges object
}

