% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{makeGtfAttributes}
\alias{makeGtfAttributes}
\title{Helper function to generate the gtf attributes field}
\usage{
makeGtfAttributes(df, cols = NULL)
}
\arguments{
\item{df}{data.frame with attribute information}

\item{cols}{character vector with column names of the attributes to be used.
With the default value of NULL all columns will be used.}
}
\description{
Helper function to generate the gtf attributes field
}

