% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_peaks.R
\name{merge.peaks}
\alias{merge.peaks}
\title{Merge MACS peaks calls from files in a directory}
\usage{
\method{merge}{peaks}(dir = "results/current/macs", FDR = 5,
  pattern = "peaks.xls$")
}
\arguments{
\item{dir}{path to the directory that holds the peak call files}

\item{FDR}{threshold for the false discovery rate}

\item{pattern}{regular expression to identify the peak files}
}
\value{
a GRanges object with the merged peak regions
}
\description{
Merge MACS peaks calls from files in a directory
}

