% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plot.spp.qc}
\alias{plot.spp.qc}
\title{Plot SPP quality metrics}
\usage{
plot.spp.qc(qc.table)
}
\arguments{
\item{qc.table}{data.frame created with \code{\link{spp.qc.summary}}}
}
\description{
Plot SPP quality metrics
}

