% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqQTL-package.R
\docType{package}
\name{seqQTL}
\alias{seqQTL}
\alias{seqQTL-package}
\title{Package for preprocessing sequencing based traits for QTL analysis}
\description{
Package for preprocessing sequencing based traits for QTL analysis
}
\details{
\tabular{ll}{
Package: \tab seqQTL\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-11-03\cr
License: \tab LGPL \cr
LazyLoad: \tab yes\cr
}
}
\examples{
# insertexample()
}
\references{
TBD
}
\keyword{package}

