% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{spp.qc.summary}
\alias{spp.qc.summary}
\title{Summary of SPP quality metrics from files in a directory}
\usage{
spp.qc.summary(dir, pattern)
}
\arguments{
\item{dir}{path to directory that holds the SPP quality results}

\item{pattern}{regular expression to identify the SPP files}
}
\value{
a table of SPP quality metrics
}
\description{
Summary of SPP quality metrics from files in a directory
}

